scripts/convert-cottereau-svgs/requirements-svg.txt
svg.path>=6.0
